Add rendering of graphs for each task.
This fixes the results page JS to properly load and render the graphs.

This also adds some missing metadata to the `/results` json response.
To get this metadata (game title, x-axis label, etc.), we modify the
game controller to (hackishly) update the game router object with a
property under a special `_swarm_` object.